+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 1. [#4268](https://github.com/influxdata/chronograf/pull/4268): Clear logs after searching
 1.  [#4253](https://github.com/influxdata/chronograf/pull/4253): Add search expression highlighting to log lines
 
+
+### UI Improvements
+1.  [#4236](https://github.com/influxdata/chronograf/pull/4236): Add spinner when loading logs table rows
+
 ### Bug Fixes
 
 1. [#4272](https://github.com/influxdata/chronograf/pull/4272): Fix logs loading description not displaying